requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
cryptography==41.0.7
python-dateutil==2.8.2
